% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos_ellipse_covariance.R
\name{argos_diag_to_cov}
\alias{argos_diag_to_cov}
\title{Transform Argos diagnostic data to covariance matrix form}
\usage{
argos_diag_to_cov(Major, Minor, Orientation)
}
\arguments{
\item{Major}{A vector containing the major axis information for each observation (na values are ok)}

\item{Minor}{A vector containing the minor axis information for each observation (na values are ok)}

\item{Orientation}{A vector containing the angle orientation of the Major axis from North (na values are ok)}
}
\value{
A \code{data.frame} with the following columns
\item{cov.x.x}{The variation of the location error in the x coordinate}
\item{cov.y.y}{The variation of the location error in the x coordinate}
\item{cov.x.y}{The covariance of the bivariate location error}
\item{cov.check}{Logical indicator that the covariance matrix is proper}
}
\description{
Using this function the user can transform the Argos diagnostic data for location
error into a form usable as a covariance matrix to approximate the location error with a
bivariate Gaussian distribution. The resulting data.frame should be attached back to the data
with \code{cbind} to use with the \code{crwMLE} function.
}
\author{
Devin S. Johnson
}
