% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_funs.R
\name{bc_raster}
\alias{bc_raster}
\title{Bhattacharyya distance for 2 raster probability densities}
\usage{
bc_raster(r, s)
}
\arguments{
\item{r}{SpatRaster object}

\item{s}{SpatRaster object}
}
\description{
Bhattacharyya distance for 2 raster probability densities
}
\details{
Computes Bhattacharyya distance for 2 raster UDs.
}
