% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clip.R
\name{find_clip}
\alias{find_clip}
\title{Find Clip Value}
\usage{
find_clip(f, prob = 0.9, upper = 1000)
}
\arguments{
\item{f}{Fraction of time unit in processed telemetry data}

\item{prob}{Cummulative probability of residency time}

\item{upper}{The maximum value for the clip value}
}
\description{
This function allows the user to calculate the clipping value for the CTMC rate matrix Q,
such that Prob(residency in a cell < (time unit)/f) = prob at the maximum movement rate. This allows
a definition, or sorts, for non-habitat. That is, if an animal would move so fast through a cell at to almost certainly
reside in a cell less than (time unit)/f length of time. Then it is essentially not habitat for the
individual.
}
