% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk-package.R
\name{gen_invLogit}
\alias{gen_invLogit}
\title{General inverse logit function}
\usage{
gen_invLogit(x, L = 0, U = 1)
}
\arguments{
\item{x}{numeric value}

\item{L}{Lower bound}

\item{U}{Upper bound}
}
\description{
General inverse logit function
}
