% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results.R
\name{get_betas}
\alias{get_betas}
\title{Summarize beta parameter values}
\usage{
get_betas(par, V = NULL, data_list)
}
\arguments{
\item{par}{Model estimated parameters}

\item{V}{Optional variance-covariance matrix for the parameters. If none is given
no standard errors will be provided for the real parameter estimates}

\item{data_list}{The data list from a fitted model object}
}
\description{
Takes fitted model and produces a summary of the beta estimates.
}
\author{
Devin S. Johnson
}
