% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results.R
\name{get_reals}
\alias{get_reals}
\title{Derive real parameter values}
\usage{
get_reals(par, V = NULL, data_list, ddl, model_parameters)
}
\arguments{
\item{par}{Model estimated parameters}

\item{V}{Optional variance-covariance matrix for the parameters. If none is given
no standard errors will be provided for the real parameter estimates}

\item{data_list}{The data list from a fitted model object}

\item{ddl}{The design data list fed into a \code{fit_mmpp} function.}

\item{model_parameters}{The model formula for the fitted model.}
}
\description{
Takes fitted model and data object and produces estimates of the
real parameter values, i.e., movement and detection rates
}
\author{
Devin S. Johnson
}
