% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Q.R
\name{load_Q}
\alias{load_Q}
\title{Load the rate matrix for the CTMC movelment model}
\usage{
load_Q(Xr, Xm, beta_r, beta_m, nb_idx, separable)
}
\arguments{
\item{Xr}{The design matrix for the residency only covariates}

\item{Xm}{The design matrix for the movement covariates}

\item{beta_r}{Coefficient vector for the residency only covariates}

\item{beta_m}{Coefficent vector for the movement covariates}

\item{nb_idx}{A 2-column vector of indecies for nonzero elements of the rate matrix.}

\item{separable}{Logical. Should the separable parameterization of Hewitt et al. (2023)
be used. Defaults to \code{separable = TRUE}}

\item{ns}{Number of sites in the study area.}
}
\description{
Load the rate matrix for the CTMC movelment model
}
\references{
Hewitt et al. (2023) xxx. Annals of Applied Statistics.
}
