% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_data.R
\name{make_design_data}
\alias{make_design_data}
\title{Produce design data for use in fitting MMPP movement models}
\usage{
make_design_data(
  cell_data,
  grad = NULL,
  rast_mask = NULL,
  directions = "rook",
  debug = 0,
  ...
)
}
\arguments{
\item{cell_data}{---.}

\item{grad}{---.}

\item{rast_mask}{Raster mask for inaccessible cells when \code{cell_data} is of type \code{SpatRaster} from the \code{terra} package. This is ignored
if \code{cell_data} is an \code{POLYGON} data frame from the \code{sf} package.}

\item{directions}{---.}

\item{debug}{Debugging level: 1-3 mainly for package developers.}

\item{...}{Ignored arguments.}
}
\description{
Produce design data for use in fitting MMPP movement models
}
