% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subset_neigh.R
\name{make_subset_neigh}
\alias{make_subset_neigh}
\title{Create neighborhood indices for subsetting cells between movement locations to improve
computational speed at the cost of some likelihood approximation.}
\usage{
make_subset_neigh(walk_data, vmax = NULL, subset_buffer = 0)
}
\arguments{
\item{walk_data}{A \code{walk_ddl} object created by \code{walk::make_walk_data}}

\item{vmax}{A single numerical value describing a rough estimate of the maximum velocity of the animal. This will be used for
for calculating the buffer around movement subsets for each time step, \verb{buffer[i] = vmax * dt[i}, for \verb{i = 1...,N}.}

\item{subset_buffer}{A value > 0 such that the bounding box of cells created
by union of the location uncertainty layers (\code{walk_data$L}) are buffered by \code{subset_buffer*walk_data$times$dt[i]} to reduce the likelihood
approximation. Once \code{subset_buffer} is large enough such that all cells are involved
at each time step, there is no longer any approximation.}
}
\description{
Create neighborhood indices for subsetting cells between movement locations to improve
computational speed at the cost of some likelihood approximation.
}
