% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ctmc.R
\name{predict_ctmc}
\alias{predict_ctmc}
\title{Predict Movement Path From Fitted CTMC Model}
\usage{
predict_ctmc(
  fit,
  walk_data,
  aux_timestamp = NULL,
  trunc_tol = 1e-08,
  debug = 0,
  ...
)
}
\arguments{
\item{fit}{A fitted model object produced by \code{\link[walk]{fit_ctmc}}}

\item{walk_data}{A design data list produced by the function \code{\link{make_walk_data}}.}

\item{aux_timestamp}{Additional POSIX times for which location prediction is desired.}

\item{trunc_tol}{Level for truncating small location prediction probabilities.}

\item{debug}{Developer debugging.}

\item{...}{Aditional arguments passed to internal function \code{ctmc_predict_arma}.}
}
\description{
Predict Movement Path From Fitted CTMC Model
}
\references{
Hewitt, J., Gelfand, A. E., & Schick, R. S. (2023). Time-discretization approximation enriches continuous-time discrete-space models for animal movement. The Annals of Applied Statistics, 17:740-760.
}
\author{
Devin S. Johnson
}
