% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_simLang.R
\name{proc_simLangevin}
\alias{proc_simLangevin}
\title{Convert \code{simLangevin}  object to a \code{SpatRaster} stack from the \code{terra} package.}
\usage{
proc_simLangevin(
  data,
  cell_data,
  return_type = "sparse",
  max_err = NULL,
  trunc = 1e-08,
  return_inadmissible = FALSE
)
}
\arguments{
\item{data}{A \code{simLangevin} object from the \code{langevinSSM} package}

\item{cell_data}{A \verb{\link[terra]\{SpatRaster\}} stack of covariates that will be used in CTMC
movement modeling. Cells with \code{NA} values will be considered areas to the animal cannot travel, i.e.,
likelihood surfaces will be \code{0} for those cells.}

\item{return_type}{Type of object returned. One if \code{"data.frame"}, \code{"sparse"} (sparse matrix),
\code{"matrix_df"} (matrix form of \code{"data.frame"}), or \code{"dense"} (dense matrix).}

\item{max_err}{The maximum error in meters. If unspecified it will be set to
4 times the maximum error standard deviation as determined by the UERE and HDOP
of the telemetry data.}

\item{trunc}{The smallest probability value that is considered to be > 0. Defaults to 1.0e-8.}

\item{return_inadmissible}{Logical. If there are locations outside the admissible areas of movement return the observations and raster cell numbers.}
}
\description{
Convert \code{simLangevin}  object to a \code{SpatRaster} stack from the \code{terra} package.
}
\details{
This function takes the HDOP information in the \code{telemetry} object to
produce a \code{SpatRaster} likelihood surface over the \code{SpatRaster} defined by
the \code{raster} argument for each location. This can then be passed to
CTMC HMM fitting functions.
}
\author{
Devin S. Johnson
}
