% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_barrier_mask.R
\name{set_barrier_mask}
\alias{set_barrier_mask}
\title{Denote cells where the animal cannot travel}
\usage{
set_barrier_mask(x, y, overlap = 0.95)
}
\arguments{
\item{x}{a polygon object that denotes places where the animal cannot travel}

\item{y}{A \code{SpatRaster} where cells will be set to \code{NA} that are inside the polygon object.}

\item{overlap}{The amount of overlap betwen the raster cell and polygon such that the cell is considered
"inside" the polygon. Defaults to \code{overlap = 0.95}. If \code{overlap = 1} then a cell has to be completely inside the
polygon to be masked.}
}
\description{
Denote cells where the animal cannot travel
}
