% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk-package.R
\name{softPlus}
\alias{softPlus}
\title{Soft-plus function}
\usage{
softPlus(x, a = 1)
}
\arguments{
\item{x}{numeric value}

\item{a}{scale parameter, \code{a} must be >1.}
}
\description{
Soft-plus function
}
