% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_Q_df.R
\name{sparse_Q_df}
\alias{sparse_Q_df}
\title{Produce design data for use in fitting MMPP movement models}
\usage{
sparse_Q_df(ras, grad = NULL, directions = "rook", debug = 0, ...)
}
\arguments{
\item{ras}{A \verb{\link[terra]\{SpatRaster\}} stack of covariate rasters that will be used in CTMC
movement modeling. Cells with \code{NA} values will be considered areas to the animal cannot travel, i.e.,
movement into those cells is not allowed. See the \verb{\link[terra]\{mask\}} function to accomplish masking of
areas where the animal cannot go.}

\item{debug}{Debugging level: 1-3 mainly for package developers.}

\item{...}{Ignored arguments.}
}
\description{
Produce design data for use in fitting MMPP movement models
}
