% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telem_to_ras.R
\name{telem_to_sparse}
\alias{telem_to_sparse}
\title{Convert {ctmm} telemtry data object to a SpatRaster stack}
\usage{
telem_to_sparse(
  data,
  cov,
  mask = NULL,
  return_type = "sparse",
  max_err = NULL,
  trunc = 1e-08
)
}
\arguments{
\item{data}{A \code{telemetry} object from the {ctmm} package}

\item{cov}{A \verb{\link[terra]\{SpatRaster\}} stack of covariates that will be used in CTMC
movement modeling.}

\item{mask}{A \verb{\link[terra]\{SpatRaster\}} object. Cells with \code{0} in the raster will be interpreted
as areas where the animal cannot go. So, likelihood surfaces will be \code{0} for those cells.}

\item{return_type}{Type of object returned. One if \code{"data.frame"}, \code{"sparse"} (sparse matrix),
\code{"matrix_df"} (matrix form of \code{"data.frame"}), or \code{"dense"} (dense matrix).}

\item{max_err}{The maximum error in meters. If unspecified it will be set to
4 times the maximum error standard deviation as determined by the UERE and HDOP
of the telemetry data.}
}
\description{
Convert {ctmm} telemtry data object to a SpatRaster stack
}
\details{
This function takes the HDOP information in the \code{telemetry} object to
produce a \code{SpatRaster} likelihood surface over the \code{SpatRaster} defined by
the \code{raster} argument for each location. This can then be passed to
CTMC HMM fitting functions.
}
\author{
Devin S. Johnson
}
