% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_model.R
\name{arg_funs}
\alias{arg_funs}
\alias{ctmc_model}
\alias{ctmc_control}
\title{Define specifics for model structure and fitting control}
\usage{
ctmc_model(form = ~1, link = "soft_plus", a = 1)

ctmc_control(
  q_r = ctmc_model(),
  q_m = ctmc_model(),
  p = FALSE,
  delta = NULL,
  form = "mult",
  norm = TRUE
)
}
\arguments{
\item{form}{Model formula for CTMC model components, residency or movement}

\item{link}{A link funtion for the residency or movement model. One of \code{"soft_plus"} or \code{"log"}.}

\item{a}{Scale parameter for the \code{"soft_plus"} link function. Ignored for \code{link="log"}.}
}
\description{
Define specifics for model structure and fitting control
}
