% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_model.R
\name{arg_funs}
\alias{arg_funs}
\alias{ctmc_model}
\alias{ctmc_control}
\title{Define specifics for model structure and fitting control}
\usage{
ctmc_model(form = ~1, link = "soft_plus", a = 1, L = 0, U = 0)

ctmc_control(
  q_r = ctmc_model(),
  q_m = ctmc_model(),
  p = FALSE,
  delta = "uniform",
  form = "mult",
  norm = TRUE,
  clip = 0
)
}
\arguments{
\item{form}{The form of the rate matrix entries. Can be one of: \code{"mult"} for residency times movement, i.e., q_{ij} = q_r(i) * pi_m(i,j), \code{"add"} for a model of
the form q_{ij} = q_r(i) + pi_m(i,j), and \code{"sde"} for a Langevin diffusion approximation. The \code{"sde"} is very similar to the additive model.}

\item{link}{A link funtion for the residency or movement model. One of \code{"soft_plus"} or \code{"log"}.}

\item{a}{Scale parameter for the \code{"soft_plus"} link function. Ignored for \code{link="log"}. The \code{a} parameter determines the approximation to
a hard plus function, i.e., as \code{a} becomes large the soft plus function converges to \code{g^{-1}(x) = max(0,x)}. For this specification, \code{a} must be greater than or equal to 1.}

\item{q_r}{A named list for the residency model. Must contain elements: \code{form}, \code{link}, and \code{a}.
The easiest way to construct this is the \code{ctmc_model} function.}

\item{q_m}{A named list for the movement portion of the model. See \code{q_r}.}

\item{p}{Logical. Should a zero-inflation parameter be used. Necessary when there might be outliers when the location likelihood does not
overlap likely locations.}

\item{delta}{Vector of probabilities for the initial location. Can be one of \code{"uniform"} for a uniform distribution over all cells, \code{"stationary"} to use the
stationary distribution implied by the parameters, or a numeric vector the same length as the number of cells.}

\item{norm}{Should the movement portion be adjusted to sum to 1. This is the parameterization suggested by Hewitt et al. (2023).}

\item{clip}{Positive value for the maximum movement rate for all cells. All cells with modeled total movement rates above this value will be truncated to \code{clip}.}
}
\description{
Define specifics for model structure and fitting control
}
\references{
Hewitt, J., Gelfand, A. E., & Schick, R. S. (2023). Time-discretization approximation enriches continuous-time discrete-space models for animal movement. The Annals of Applied Statistics, 17:740-760.
}
