% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_likelihood.R
\name{ctmc_n2ll}
\alias{ctmc_n2ll}
\title{Evaluate movement MMPP log-likelihood}
\usage{
ctmc_n2ll(par, data_list, debug = 0, ...)
}
\arguments{
\item{par}{Parameter vector}

\item{data_list}{List of required data objects to evaluate likelihood}

\item{debug}{For developers only, leave in the default setting.}

\item{...}{Extra wiggle room for ignored arguments.}
}
\description{
Evaluate movement MMPP log-likelihood
}
\author{
Devin S. Johnson
}
