% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ctmc.R
\name{fit_ctmc}
\alias{fit_ctmc}
\title{Fit CTMC movement model}
\usage{
fit_ctmc(
  Q_dd,
  L,
  model_parameters = list(Q = list(res_form = ~1, move_form = ~1, separable = TRUE), L =
    NULL),
  hessian = TRUE,
  start = NULL,
  method = "nlminb",
  fit = TRUE,
  debug = 0,
  ...
)
}
\arguments{
\item{Q_dd}{Design data produced by the function \code{\link{sparse_Q_df}}.}

\item{model_parameters}{Model formula for the detection and movement portions
of the MMPP model. Must be of the form, e.g.,
\code{list(lambda=list(form=~1, offset=NULL), q=list(form=~1, offset=~log(1.num_neigh)))}. For the
\code{q} model you must use \code{offset=0} to not have one. If it is left off, \code{offset=~log(1.num_neigh)))}
will be used.}

\item{hessian}{Logical. Should the Hessian matrix be calculated to obtain the parameter
variance-covariance matrix.}

\item{start}{Optional starting values for the parameter must be a list of the
form \code{list(beta_l=c(), beta_q=c())}.}

\item{method}{Optimization method. See \code{\link[optimx]{optimr}}}

\item{fit}{Logical. Should the likelihood be optimized?}

\item{debug}{Integer from 1-4. Opens browser() at various points in the function call. Mostly for
package developers.}

\item{...}{Additional arguments passed to the optimization function
\code{\link[optimx]{optimr}} from the \code{\link[optimx]{optimx-package}}.}

\item{Lik_mat}{Sparse matrix}
}
\description{
Fit CTMC movement model
}
\author{
Devin S. Johnson
}
