% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ctmc.R
\name{fit_ctmc}
\alias{fit_ctmc}
\title{Fit CTMC movement model}
\usage{
fit_ctmc(
  Q_dd,
  Lik_mat,
  model_parameters = list(Q = list(q_r = ~1, q_m = ~1, separable = TRUE), L = NULL),
  hessian = TRUE,
  start = NULL,
  method = "nlminb",
  fit = TRUE,
  debug = 0,
  ...
)
}
\arguments{
\item{Q_dd}{Design data list produced by the function \code{\link{make_Q_data}}.}

\item{Lik_mat}{Sparse matrix where each row is the likelihood surface for the corresponding location observation.x}

\item{model_parameters}{Model formula for the residency and movement portions of the model, e.g.,
\code{list(Q = list(q_r=~1, q_m=~1, separable=TRUE), L = NULL)}.}

\item{hessian}{Logical. Should the Hessian matrix be calculated to obtain the parameter
variance-covariance matrix.}

\item{start}{Optional starting values for the parameter must be a list of the
form \code{list(beta_l=c(), beta_q=c())}.}

\item{method}{Optimization method. See \code{\link[optimx]{optimr}}}

\item{fit}{Logical. Should the likelihood be optimized?}

\item{debug}{Integer from 1-4. Opens browser() at various points in the function call. Mostly for
package developers.}

\item{...}{Additional arguments passed to the optimization function
\code{\link[optimx]{optimr}} from the \code{\link[optimx]{optimx-package}}.}
}
\description{
Fit CTMC movement model
}
\details{
The \code{separable = TRUE} element of the \code{Q} list for \code{model.parameters} indicates that the
separable parameteriztion of Hewitt et al. (2023) will be used. If set to \code{separable = FALSE}
the traditional loglinear formulation of Johnson et al. (2021) and Hanks et al. (2015) will be used. If  \code{separable = FALSE} the
\code{q_r} formula is ignored.
}
\references{
Hanks, E. M., Hooten, M. B., and Alldredge, M. W. (2015) Continuous-time discrete-space models for animal movement. Annals of Applied Statistics. 9:145-165.

Hewitt, J., Gelfand, A. E., & Schick, R. S. (2023). Time-discretization approximation enriches continuous-time discrete-space models for animal movement. The Annals of Applied Statistics, 17:740-760.

Johnson, D. S., Pelland, N. A., and Sterling, J. T. (2021) A Continuous-Time Semi-Markov Model for Animal Movement in a Dynamic Environment. The Annals of Applied Statistics, 15:797-812.
}
\author{
Devin S. Johnson
}
