% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_functions.R
\name{get_lim_ud}
\alias{get_lim_ud}
\title{Get the limiting utilization distribution of the CTMC movement process}
\usage{
get_lim_ud(fit = NULL, hpd = NULL, method = "lu", ...)
}
\arguments{
\item{fit}{A fitted model object from \code{\link[walk]{fit_ctmc}}.}

\item{hpd}{A vector of probabilities. Will return columns with highest probability area for each specified probability. E.g.,
\code{hpd=c(0.5, 0.95)} will return 2 extra columns with 50 and 95\% HPD densities.}

\item{method}{Method used for eigen decomposition. One of \code{"lu"} or \code{"arpack"}.}

\item{...}{Extra arguments to pass to \code{\link[rARPACK]{eigs}}}
}
\description{
Get the limiting utilization distribution of the CTMC movement process
}
\author{
Devin S. Johnson
}
