% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk-package.R
\docType{package}
\name{walk-package}
\alias{walk}
\alias{walk-package}
\title{walk: Model Animal Movement With Continuous-Time Markov Chains}
\description{
Model animal movement using continuous-time Markov chain models.
}
\author{
\strong{Maintainer}: Devin S. Johnson \email{devin.johnson@noaa.gov}

}
\keyword{internal}
