% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk-package.R
\docType{package}
\name{walk-package}
\alias{walk}
\alias{walk-package}
\title{walk: What the Package Does (One Line, Title Case)}
\description{
Model animal movement using continuous-time Makov chain models.
}
\author{
\strong{Maintainer}: Devin Johnson \email{devin.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-0068-7098}{ORCID})

}
\keyword{internal}
