% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk-package.R
\docType{package}
\name{walk-package}
\alias{walk}
\alias{walk-package}
\title{walk: Model Animal Movement With Continuous-Time Makov Chains}
\description{
Model animal movement using continuous-time Makov chain models.
}
\author{
\strong{Maintainer}: Devin S. Johnson \email{devin.johnson@noaa.gov}

Authors:
\itemize{
  \item James T. Thorson \email{james.throson@noaa.gov}
  \item Janelle J. Badger \email{janelle.badger@noaa.gov}
}

}
\keyword{internal}
